Add an usage example (#54)
Great project everyone! I appreciate it a lot.

My only issue was that I could not find an example on how to actually use the api. I had to look up in the source code that the endpoint is `/api/icd10` with query parameter `search`. I think people would appreciate a simple example in the README to get going faster :)

Co-authored-by: andreakreichgauer <38720222+andreakreichgauer@users.noreply.github.com>