feat: return only complete codesystem concepts
# Ticket 🎫

This closes #60.

# Description 📖
This filters the search results for presence of the properties `code` and `display` and logs a warning if incomplete concepts are found.

# How to Test? 🧪

Remove a code or a display property from the ICD-10 CodeSystem and search for this code. It should not appear in the search results anymore. A warning should be logged by the API.

# Checklist ✅
- [x] My changes generate no new warnings
- [x] Made corresponding changes to the documentation if applicable
- [x] Commented my code, particularly in hard-to-understand areas if applicable
- [x] Performed a self-review of my own changes